# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KYD marketplace API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type CaregiverServicesItem =
  (typeof CaregiverServicesItem)[keyof typeof CaregiverServicesItem];

export const CaregiverServicesItem = {
  baby: "baby",
  pet: "pet",
  elder: "elder",
} as const;

export interface Caregiver {
  id: string;
  name: string;
  city: string;
  bio: string;
  services: CaregiverServicesItem[];
  pricePerHour: number;
  rating: number;
  reviewCount: number;
  availableNow: boolean;
  verified: boolean;
  avatarColor: string;
  lat: number;
  lng: number;
  radiusKm: number;
  certifications: string[];
  yearsExperience: number;
}

export type AvailabilitySlotSlotsItem = {
  hour: string;
  free: boolean;
};

export interface AvailabilitySlot {
  day: string;
  slots: AvailabilitySlotSlotsItem[];
}

export type CaregiverDetail = Caregiver & {
  availability: AvailabilitySlot[];
};

export interface Review {
  id: string;
  caregiverId: string;
  caregiverName: string;
  authorName: string;
  authorRole: string;
  city: string;
  rating: number;
  comment: string;
  date: string;
  verified: boolean;
}

export interface Booking {
  id: string;
  caregiverId: string;
  caregiverName: string;
  familyName: string;
  service: string;
  date: string;
  hours: number;
  total: number;
  status: string;
  createdAt: string;
}

export interface CreateBookingInput {
  caregiverId: string;
  familyName: string;
  service: string;
  date: string;
  hours: number;
  notes?: string;
}

export interface ChatMessage {
  id: string;
  room: string;
  author: string;
  text: string;
  createdAt: string;
}

export interface SendChatMessageInput {
  author: string;
  text: string;
}

export interface StatsOverview {
  families: number;
  caregivers: number;
  reviews: number;
  successRate: number;
  cities: number;
  distributedEur: number;
  avgResponseMin: number;
  treesPlanted: number;
}

export interface CityStat {
  city: string;
  caregivers: number;
}

export type ListCaregiversParams = {
  service?: ListCaregiversService;
  city?: string;
  maxPrice?: number;
  availableNow?: boolean;
};

export type ListCaregiversService =
  (typeof ListCaregiversService)[keyof typeof ListCaregiversService];

export const ListCaregiversService = {
  baby: "baby",
  pet: "pet",
  elder: "elder",
  all: "all",
} as const;