# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: KYD marketplace API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
  - name: caregivers
  - name: reviews
  - name: bookings
  - name: chat
  - name: stats
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /caregivers:
    get:
      operationId: listCaregivers
      tags: [caregivers]
      summary: List caregivers with optional filters
      parameters:
        - in: query
          name: service
          schema:
            type: string
            enum: [baby, pet, elder, all]
        - in: query
          name: city
          schema:
            type: string
        - in: query
          name: maxPrice
          schema:
            type: number
        - in: query
          name: availableNow
          schema:
            type: boolean
      responses:
        "200":
          description: List of caregivers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Caregiver"
  /caregivers/{id}:
    get:
      operationId: getCaregiver
      tags: [caregivers]
      summary: Get caregiver details
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Caregiver detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaregiverDetail"
        "404":
          description: Not found
  /caregivers/{id}/reviews:
    get:
      operationId: listCaregiverReviews
      tags: [reviews]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Reviews for a caregiver
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
  /reviews/featured:
    get:
      operationId: listFeaturedReviews
      tags: [reviews]
      summary: Top reviews to feature on the homepage
      responses:
        "200":
          description: Featured reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
  /bookings:
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingInput"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/recent:
    get:
      operationId: listRecentBookings
      tags: [bookings]
      summary: Recent bookings activity
      responses:
        "200":
          description: Recent bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
  /chat/{room}/messages:
    get:
      operationId: listChatMessages
      tags: [chat]
      parameters:
        - in: path
          name: room
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
    post:
      operationId: sendChatMessage
      tags: [chat]
      parameters:
        - in: path
          name: room
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendChatMessageInput"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessage"
  /stats/overview:
    get:
      operationId: getStatsOverview
      tags: [stats]
      summary: Marketplace stats counters
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatsOverview"
  /stats/cities:
    get:
      operationId: listTopCities
      tags: [stats]
      summary: Top cities by caregiver count
      responses:
        "200":
          description: Cities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CityStat"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]
    Caregiver:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        city:
          type: string
        bio:
          type: string
        services:
          type: array
          items:
            type: string
            enum: [baby, pet, elder]
        pricePerHour:
          type: number
        rating:
          type: number
        reviewCount:
          type: integer
        availableNow:
          type: boolean
        verified:
          type: boolean
        avatarColor:
          type: string
        lat:
          type: number
        lng:
          type: number
        radiusKm:
          type: number
        certifications:
          type: array
          items:
            type: string
        yearsExperience:
          type: integer
      required: [id, name, city, bio, services, pricePerHour, rating, reviewCount, availableNow, verified, avatarColor, lat, lng, radiusKm, certifications, yearsExperience]
    CaregiverDetail:
      allOf:
        - $ref: "#/components/schemas/Caregiver"
        - type: object
          properties:
            availability:
              type: array
              items:
                $ref: "#/components/schemas/AvailabilitySlot"
          required: [availability]
    AvailabilitySlot:
      type: object
      properties:
        day:
          type: string
        slots:
          type: array
          items:
            type: object
            properties:
              hour:
                type: string
              free:
                type: boolean
            required: [hour, free]
      required: [day, slots]
    Review:
      type: object
      properties:
        id:
          type: string
        caregiverId:
          type: string
        caregiverName:
          type: string
        authorName:
          type: string
        authorRole:
          type: string
        city:
          type: string
        rating:
          type: integer
        comment:
          type: string
        date:
          type: string
        verified:
          type: boolean
      required: [id, caregiverId, caregiverName, authorName, authorRole, city, rating, comment, date, verified]
    Booking:
      type: object
      properties:
        id:
          type: string
        caregiverId:
          type: string
        caregiverName:
          type: string
        familyName:
          type: string
        service:
          type: string
        date:
          type: string
        hours:
          type: number
        total:
          type: number
        status:
          type: string
        createdAt:
          type: string
      required: [id, caregiverId, caregiverName, familyName, service, date, hours, total, status, createdAt]
    CreateBookingInput:
      type: object
      properties:
        caregiverId:
          type: string
        familyName:
          type: string
        service:
          type: string
        date:
          type: string
        hours:
          type: number
        notes:
          type: string
      required: [caregiverId, familyName, service, date, hours]
    ChatMessage:
      type: object
      properties:
        id:
          type: string
        room:
          type: string
        author:
          type: string
        text:
          type: string
        createdAt:
          type: string
      required: [id, room, author, text, createdAt]
    SendChatMessageInput:
      type: object
      properties:
        author:
          type: string
        text:
          type: string
      required: [author, text]
    StatsOverview:
      type: object
      properties:
        families:
          type: integer
        caregivers:
          type: integer
        reviews:
          type: integer
        successRate:
          type: integer
        cities:
          type: integer
        distributedEur:
          type: integer
        avgResponseMin:
          type: integer
        treesPlanted:
          type: integer
      required: [families, caregivers, reviews, successRate, cities, distributedEur, avgResponseMin, treesPlanted]
    CityStat:
      type: object
      properties:
        city:
          type: string
        caregivers:
          type: integer
      required: [city, caregivers]