# lib/api-zod/src/generated/types/availabilitySlot.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KYD marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { AvailabilitySlotSlotsItem } from "./availabilitySlotSlotsItem";

export interface AvailabilitySlot {
  day: string;
  slots: AvailabilitySlotSlotsItem[];
}