# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KYD marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface Booking {
  id: string;
  caregiverId: string;
  caregiverName: string;
  familyName: string;
  service: string;
  date: string;
  hours: number;
  total: number;
  status: string;
  createdAt: string;
}