# lib/api-zod/src/generated/types/caregiver.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KYD marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { CaregiverServicesItem } from "./caregiverServicesItem";

export interface Caregiver {
  id: string;
  name: string;
  city: string;
  bio: string;
  services: CaregiverServicesItem[];
  pricePerHour: number;
  rating: number;
  reviewCount: number;
  availableNow: boolean;
  verified: boolean;
  avatarColor: string;
  lat: number;
  lng: number;
  radiusKm: number;
  certifications: string[];
  yearsExperience: number;
}