# lib/api-zod/src/generated/types/caregiverDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KYD marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { AvailabilitySlot } from "./availabilitySlot";
import type { Caregiver } from "./caregiver";

export type CaregiverDetail = Caregiver & {
  availability: AvailabilitySlot[];
};