# lib/api-zod/src/generated/types/chatMessage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KYD marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface ChatMessage {
  id: string;
  room: string;
  author: string;
  text: string;
  createdAt: string;
}