# lib/api-zod/src/generated/types/createBookingInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KYD marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateBookingInput {
  caregiverId: string;
  familyName: string;
  service: string;
  date: string;
  hours: number;
  notes?: string;
}