# lib/api-zod/src/generated/types/review.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KYD marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface Review {
  id: string;
  caregiverId: string;
  caregiverName: string;
  authorName: string;
  authorRole: string;
  city: string;
  rating: number;
  comment: string;
  date: string;
  verified: boolean;
}