# lib/api-zod/src/generated/types/sendChatMessageInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KYD marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface SendChatMessageInput {
  author: string;
  text: string;
}