# lib/api-zod/src/generated/types/statsOverview.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KYD marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface StatsOverview {
  families: number;
  caregivers: number;
  reviews: number;
  successRate: number;
  cities: number;
  distributedEur: number;
  avgResponseMin: number;
  treesPlanted: number;
}